Add star2lte device repos
Maru on the Samsung Galaxy S9+ coming soon...

This includes some project dependencies like hardware/samsung which are
not pulled in by default in the Lineage manifest: they are usually
pulled in via "breakfast" at build time for stock Lineage devices. Maru
uses the normal "lunch && make dist" in stock AOSP so we need to make
sure to pull all necessary dependencies in explicitly in the manifest.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>